[CLIENT] Splits AttributeSingle into AttributeInfoButton and AttributeDescription
- Adds id to AttributeItem
- Removes value from AttributeItem
