+++ src/client/src/content/home/types/AttributeItem.ts
@@ -1,6 +1,6 @@
 export interface AttributeItem {
+  id: string;
   name: string;
   color: string;
   traits: { [key: string]: string };
-  value: string;
 }
